Adding playbooks for overcloud backup and restore
These are playbooks designed to backup and restore an overcloud.
Currently they only support an RPM-based overcloud (e.g. ocata,
newton, etc) but work is being done to accommodate containerized
overclouds.

The backup playbook is pretty straightforward. You run the playbook
and it'll create a backup of the databases (MySQL, MongoDB, and Redis)
and then create a backup of each Controller file system. Probably the
only pre-req is that the undercloud needs SSH access to itself to
perform certain Ansible tasks on itself.

When running the restore playbook, you need to set the
overcloud_backup_tar variable to the location of the backup tar. For
example:

$ ansible-playbook -i inventory \
--ssh-common-args="-o StrictHostKeyChecking=no" \
-e "overcloud_backup_tar=overcloud-backup-20180711T121449.tar" \
~/overcloud-backup/overcloud-restore.yaml

This playbook will restore the MySQL database only. Further restore
functions will be added in the future.

Also work noting that both playbooks should work with
tripleo-ansible-inventory script and static inventories that follow
the same basic structure. For example:

[undercloud]
192.0.2.1

[Controller]
192.0.2.100
192.0.2.101
192.0.2.102

Change-Id: Ia43b4a4c2412b31b9e947f9ab3fc21fbcf2c1ca8